+++ src/components/common/Header/index.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import * as S from './style';
+
+const Header = () => {
+  return (
+    <S.HeaderWrapper>
+      <S.PageTitle>점심 뭐 먹지</S.PageTitle>
+    </S.HeaderWrapper>
+  );
+};
+
+export default Header;
