+++ src/components/common/SelectBox/style.ts
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const Select = styled.select`
+  height: 44px;
+  min-width: 125px;
+  padding: 8px;
+  border: 1px solid var(--grey-200);
+  border-radius: 8px;
+  background: transparent;
+  font-size: 16px;
+`;
