+++ src/components/restaurant/CategoryIcon/style.ts
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+
+export const CategoryIconWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 64px;
+  height: 64px;
+  min-width: 64px;
+  min-height: 64px;
+  margin-right: 16px;
+  border-radius: 50%;
+  background: var(--lighten-color);
+  img {
+    width: 36px;
+    height: 36px;
+  }
+`;
