+++ src/components/restaurant/RestaurantList/style.ts
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const RestaurantListContainer = styled.main`
+  padding: 16px;
+`;
+
+export const SelectBoxContainer = styled.div`
+  display: flex;
+  justify-content: space-between;
+`;
+
+export const RestaurantList = styled.ul`
+  display: flex;
+  flex-direction: column;
+  margin: 16px 0;
+`;
