+++ src/style/globalStyle.ts
@@ -1,4 +1,4 @@
-import { createGlobalStyle } from 'styled-components';
+import { createGlobalStyle, css } from 'styled-components';
 
 const GlobalStyle = createGlobalStyle`
   * {
@@ -30,30 +30,30 @@ const GlobalStyle = createGlobalStyle`
     --backdrop-color: #00000059;
     --divide-color: #e9eaed;
   }
+`;
 
-  .text-title {
-    font-size: 20px;
-    line-height: 24px;
-    font-weight: 600;
-  }
+export const textTitle = css`
+  font-size: 20px;
+  line-height: 24px;
+  font-weight: 600;
+`;
 
-  .text-subtitle {
-    font-size: 18px;
-    line-height: 28px;
-    font-weight: 600;
-  }
+export const textSubTitle = css`
+  font-size: 18px;
+  line-height: 28px;
+  font-weight: 600;
+`;
 
-  .text-body {
-    font-size: 16px;
-    line-height: 24px;
-    font-weight: 400;
-  }
+export const textBody = css`
+  font-size: 16px;
+  line-height: 24px;
+  font-weight: 400;
+`;
 
-  .text-caption {
-    font-size: 14px;
-    line-height: 20px;
-    font-weight: 400;
-  }
+export const textCaption = css`
+  font-size: 14px;
+  line-height: 20px;
+  font-weight: 400;
 `;
 
 export default GlobalStyle;
